:art: refactor style code
