+++ src/components/common/datetime/DatePicker.tsx
@@ -1,3 +1,4 @@
+// Unused component
 import * as React from "react";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
